requirements.txt
datasets>=2.19.0
transformers>=4.41.0
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
pymupdf==1.24.9
beautifulsoup4==4.12.3
lxml>=5.2.2
pandas>=2.2.2
torch>=2.2.0
numpy==1.26.4
pypdf>=4.2.0
unidecode>=1.3.8


# Core Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Templates
jinja2>=3.1.4
